+++ R/googleAuthR_generator.R
@@ -282,6 +282,11 @@ retryRequest <- function(f){
 #' @family data fetching functions
 checkTokenAPI <- function(shiny_access_token=NULL){
 
+  if(any(which(grepl("with_mock_API", as.character(sys.calls()))))){
+    myMessage("Skipping token checks as using with_mock_API", level = 3)
+    return(TRUE)
+  }
+
   if(!is.null(gar_cache_get_loc())){
     myMessage("Skipping token checks as using cache", level = 3)
     return(TRUE)